+++ sqrbot/handlers/__init__.py
@@ -1,2 +1,3 @@
 from .index import *
 from .event import *
+from .interaction import *